added pagination & file upload and download

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require("express-validator");
 const Product = require("../models/Product");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
     res.render("admin/edit-product", {
@@ -14,12 +15,29 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    // const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
 
     const errors = validationResult(req);
-    console.log(description);
+
+    if (!image) {
+        return res.status(422).render("admin/edit-product", {
+            pageTitle: "Add Product",
+            path: "/admin/add-product",
+            editing: false,
+            prod: {
+                title: title,
+                price: price,
+                description: description,
+            },
+            hasError: true,
+            errorMessage: "attached file is not an image",
+        });
+    }
+
+    const imageUrl = image.path;
 
     if (!errors.isEmpty()) {
         return res.status(422).render("admin/edit-product", {
@@ -85,7 +103,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDesc = req.body.description;
 
@@ -96,7 +114,10 @@ exports.postEditProduct = (req, res, next) => {
             }
             product.title = updatedTitle;
             product.price = updatedPrice;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = updatedImageUrl;
+            }
             product.description = updatedDesc;
             return product.save().then((result) => {
                 console.log("updated Product!");
@@ -109,14 +130,21 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.deleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    const prodId = req.params.productId;
+    Product.findById(prodId)
+        .then((product) => {
+            if (!product) {
+                return next(new Error("Product not found!"));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(() => {
             console.log("deleted products");
-            res.redirect("/admin/products");
+            res.status(200).json({ message: "success!" });
         })
         .catch((err) => {
-            console.log(err);
+            res.status(500).json({ message: "Deleting product failed!" });
         });
 };
